requirements.txt
python-arango==4.4.0
